+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completions for `--class` and `-t` (short title)
 - Change the mouse cursor when hovering over the message bar and its close button
 - Support combined bold and italic text (with `font.bold_italic` to customize it)
-- Extra bindings for F13-F24
+- Extra bindings for F13-F20
+- Bindings with combined modifiers
 - Bindings for ScrollToTop and ScrollToBottom actions
 
 ### Changed
@@ -47,7 +48,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wayland primary selection clipboard not storing text when selection is stopped outside of the window
 - Block URL highlight while a selection is active
 - Bindings for Alt + F1-F12
-- Missing bindings for combined modifiers
 
 ### Removed